properly handle organ selection
- make sure what user picks is supported.
- properly set the organ based on user's previous selection
